Added support for /docs/latest
