+++ .gitignore
@@ -66,3 +66,4 @@ build/
 src/website/data/docs.yml
 src/website/data/ecosystem.yml
 src/website/content/docs/*/*.md
+.DS_Store
