+++ src/website/layouts/partials/head.html
@@ -29,6 +29,10 @@
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" integrity="sha256-eZrrJcwDc/3uDhsdt61sL2oOBY362qM3lon1gyExkL0=" crossorigin="anonymous" />
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.12.0/styles/default.min.css" integrity="sha256-Zd1icfZ72UBmsId/mUcagrmN7IN5Qkrvh75ICHIQVTk=" crossorigin="anonymous" />
 
+  {% if canonical %}
+    <link ref="canonical" href="{{ canonical }}">
+  {% endif %}
+
   {% if google_analytics %}
     {% include 'partials/google-analytics.html' %}
   {% endif %}
