+++ src/website/layouts/partials/header.html
@@ -20,7 +20,7 @@
   <div class="navbar-menu" id="mobileMenu">
     <div class="navbar-start">
       <a class="navbar-item " href="/">Home</a>
-      <a class="navbar-item " href="{{ data.docs.toc[data.docs.versions[1]][0].link }}">Docs</a>
+      <a class="navbar-item " href="/docs/latest">Docs</a>
       <a class="navbar-item " href="/ecosystem">Ecosystem</a>
       <a class="navbar-item " href="/benchmarks">Benchmarks</a>
     </div>
